+++ src/textual/screen.py
@@ -1278,7 +1278,7 @@ class Screen(Generic[ScreenResultType], Widget):
                 tooltip.display = False
             else:
                 tooltip.display = True
-                tooltip._absolute_offset = self.app.mouse_position
+                tooltip.absolute_offset = self.app.mouse_position
                 tooltip.update(tooltip_content)
 
     def _handle_mouse_move(self, event: events.MouseMove) -> None: